[#782] Move observe utility into common and move sqldelight closable
