+++ ios/CCC/Util/ObservableSEED.swift
@@ -26,7 +26,7 @@ final class ObservableSEED<
 
     let data: Data?
 
-    private var closeable: Ktor_ioCloseable!
+    private var closeable: RuntimeCloseable!
 
     // swiftlint:disable force_cast
     init(viewModel: ViewModel) {
@@ -46,12 +46,12 @@ final class ObservableSEED<
         logger.i(message: {"ObservableSEED \(ViewModel.description()) startObserving"})
 
         if viewModel.state != nil {
-            closeable = viewModel.observe(viewModel.state!, onChange: {
+            closeable = CoroutineUtilKt.observeWithCloseable(viewModel.state!, onChange: {
                 self.state = $0 as! State
             })
         }
         if viewModel.effect != nil {
-            closeable = viewModel.observe(viewModel.effect!, onChange: {
+            closeable = CoroutineUtilKt.observeWithCloseable(viewModel.effect!, onChange: {
                 self.effect.send($0 as! Effect)
             })
         }
